NDEX-676 (Filtering in network edges):
Replaced all calls to refreshEdgeTable() and refreshNodeTable() with populateEdgeTable(enableFiltering) and populateNodeTable(enableFiltering),
respectively. So, every time we run query, we now redefine the table and enable filtering.  This is not a big overhead.  I didn't managed to enable filtering by turning the enableFiltering property in the network definition to true, so for queries we redefine the table and re-populate it.

NDEX-669 (download links of exported networks open file instead of downloading):

Modified user.html; added "download" tag to
<a ng-show="task.status == 'COMPLETED' && task.taskType == 'EXPORT_NETWORK_TO_FILE'"
     href="/exported-networks/{{task.externalId}}.{{userController.getTaskFileExt(task)}}" download> Download Network
</a>