This resolves NDEX-656 (Empty Cells in Edge Property Column are not handled properly by the edge table grid view).
refreshEdgeTable and refreshNodeTabe functions parse network properties and add them, row by row, to the Edges and Nodes tables, respectively.

I modified these two functions to explicitly add empty values to the rows for the missing properties.  In other words, every
row that we add to the tables should have as many properties as there are property headers in the tables; and if a property is missing from
the table row that we construct, then we add this property and assign it an empty value.