These changes resolve NDEX-692 (adding group or user should not happen unless the save changes button is clicked)
1) added a new factory.getAccountsByUUIDs() function to Ndex-service-factory.js;  this call uses the newly-added server POST function at path /admin/accounts
that maps to the new REST API getAccountsByuuids() which returns to the client a list of Account objects.  This list of objects is parsed by the manage-network-access.js.
We need to know  1) whether an account object received from the server is user or group 2) First Name and Last Name for User accounts, and Group Name for Group Accounts.

2) modifications to the manage-network-acess.js:
- removed caching of user and group UUIDs that was done by calling networkManager.loadAllUsers(); and networkManager.loadAllGroups();
- networkManager.processReceivedAccessPermissions() now builds a list of UUIDs that have permissions for the current network, and sends this list
to the server using the new client function factory.getAccountsByUUIDs().  It then receives a list of account objects from the server, iterates through this list, checks
whether the account info received is user or group, gets First Name and Last Name for user account info, and Group Name for group accounts, creates newMembership object
and  initializes both networkManager.originalAccessPermissions and networkManager.selectedAccountsForUpdatingAccessPermissions
with this newMembership object.

To summarize, before we had to retrieve and cache all user and group account objects and iterate through them to get their user first/last or group names.
Now with the new functionality added to the server (/admin/accounts POST ), we can  retrieve account information for just the UUIDs we are interested in.