This is for NDEx-611 (Add HTML/Markdown support to "Description" and "Citation" fields in edit network profile):
1) app.js -- removed injection of 'markdown' into ndexApp since we now use textAngular library
2) removed libraries ndex-webapp/lib/angular-markdown-text-master and ndex-webapp/liob/showdown since we now use textAngular
3) index.html -- removed inclusion of showdown.min.js and angular-markdown-text.min.js since we do not use them any longer
4) in bulkEditNetworkPropertyModal.html, added ng-click directive to the Change Description, Change Version and Change Visibility buttons that set network.operation
model so that bulkEditNetworkPropertyModal could know waht type of bulk operation is performed, and act accordingly
5) in edit-network-properties-controller.js added editor.reference model that keeps Reference property after it is removed from the dis[lay in the Properties portion of the screen.
Also, added code that adds the Reference back (if it was originally present) to the properties structure before updating properties on the server.
ndexService.getNetwork() function ow removes Reference property from the Properties structure received from the server and saves it to editor.reference.
This is done because we do not want to show the Reference property in the Properties portion of the screen or in the Properties Editor.  We show it in the
top portion of the screen below network description, and we edit this property in the Edit NEtwork Profile modal.  So, when we receive the properties from the server,we remove
Reference form the properties and and it to the network structure. And when we update Properties, we add Reference back to the Properties.
6) editNetworkSummaryModal.html -- we now only have network.reference as the model, and not network.reference[1,2]
7) in ndex-angular-site.css, increased max-width at which top menu collapsed to 1210px from 1200px.
8) network-controller.js -- removed networkController.currentNetworkReference; replaced it with networkController.currentNetwork.reference.  The getNetworkReference() function now removes Reference
from the list of properties received from the server and saves it in networkController.currentNetwork.reference.
9) modified network.html to use networkController.currentNetwork.reference instead of networkController.currentNetworkReference.  Replaced the Edit Properties glyph with a link.
10) ui-services.js -- simplified the logic of editNetworkSummaryModal when checking whether Reference property was modified or not and deciding if it needs to updated on the server.
Modified bulkEditNetworkProperty to only set two fields in the property structure that is sent to the server for updating the porperty: network UUID and the property that needs to be updated (description, visibility or version).