This is for NDEx-611 (Add HTML/Markdown support to "Description" and "Citation" fields in edit network profile):
1) index.html - added to the project to-markdown.min.js (documented at https://github.com/domchristie/to-markdown).
This library contains toMarkdown() function that converts HTML to markdown.  We use markdown in pop-up windows
of network and netwpork search pages for network description.  Markdown is preferrable to simple text in these cases
since markdown still has structure (keeps new lines, paragrapks, makes heading to stand out, etc).

2) added stripHTML() function to network-search-controller.js and user-controller.js.  This function removes most HTML
tags and replaces them with markdown symbols so that this field could be displayed in the title element of networkName.html
template in the pop-up window when mouse cursor hovers over it.