This resolves NDEX-741 (implement phase 1 of TCL as per our specs in v1.3 technical docs). It implements the support of 4 TLC attributes (rightsHolder, copyrightDate, licenseType, and publicRightsType). Here is what has been done:
1) added "Licensing Info" to the network.html.  This info shown if licensing information is present.

2) in network-controller.js, added networkController.currentNetwork.licensingInfo property that is initialized by the new getLicensingInfo() function.
getLicensingInfo() traverses through the list of network properties, retrieves all rightsHolder, copyrightDate, licenseType, and publicRightsType properties
and concatenates them into a string that is presented on the Network page.

Here is how the licensing info is built:

1) all of the attributes are case-insensitive (i.e.,  publicRightsType is the same as publicrightstype or PUBLICrightsTyPE)
2) there is no limit or check on how many attributes of the same type can be  specified.  In other words, user can specify
any number of elements of the same type, and they all will be concatenated into a string with individual elements separated by space.
For example, user can specify 2 elements rightsHolder, "rh1" and "rh2", and they will be concatenated into one string "rh1 rh2" when rendered on the page.
3) the order of different types of elements when presented on the page is
 - rightsHolder
 - copyrightDate
 - licenseType
 - publicRightsType
4) the order of presentation of elements of the same type corresponds to the order in which these elements are entered in the editor.
5) all of these elements are edited in the Edit Properties (there is link to it on the Network page)