This commit resolves NDEx-740 (Reference property on Network Page).
Summary:  if user tries to add a Reference predicate property on Custom Network Properties Editor page,
(s)he is now presented with a warning modal that Reference is s reserved word (case-inseinsitive),
And therefore will not be saved. The check for Reference is performed on losing focus from the input field.

Changes:

1) editNetworkProperties.html – added onblur="checkPredicateString(this);” that calls the newly-added checkPredicateString() function.

2) edit-network-properties-controller.js - added a new checkPredicateString() function that checks if user entered “reference” in the predicate field.  If yes, a warning is shown.  Also, before sending the properties
to the server, we now remove all “reference” fields that were entered by the user.