This resolves NDEX-741 (implement phase 1 of TCL as per our specs in v1.3 technical docs)
Summary of changes:  removed support (special treatment) of copyrightDate, rightsHolder, licenseType, and publicRIghtsType.  Added support
of rightsHolder and rights as described in NDEX-741.

- in network-controller.js, removed networkController.currentNetwork.licensingInfo;  added networkController.currentNetwork.rights and
networkController.currentNetwork.rightsHolder.  Modified getLicensingInfo function to parse the list of network properties and add
rights and rightsHolder info to the correspondent arrays.  Added a new function isPredicateReservedWord that checks if the word passed as
parameter is empty, "rights" or "rightsHolder".

- modififed network.html to show rightsHolder and rights information at the top of the screen and do not show these properties in the Network Properties section.
 Removed code that showed licensingInfo.