This commit resolves NDEX-786 (clicking owner name in network search result displays the wrong networks):
1) ndex-service-factory.js - modified factory.getUserNetworkMembership() not to call successHandler() in case externalId is undefined or null
because in this case the caller starts iterating through the list of networks it expects from  getUserNetworkMembership(), and that list of networks
is not defined

2) user-controller.js - modified userController.submitNetworkSearch() to check whether the search for networks is done
a) anonymously or for networks owned by user other than the logged in one
b) to get networks the currently logged in user has at least READ access to (both owned by the loged in user and owned by other users/accounts).

Depending on the scenario (1 or 2 described above), we set the fields in networkQuery accordingly.