1) Small modification in factory.cyElementsFromNiceCX() - we should check for "nodeElement.n" not for "nodeElement.name" since niceCX contains nodeElement.n.  This value (if it exists) should be assigned to nodeData.name.
2) - in cx-network-utils.js, modified all global functions that begin with "this", to begin with "self" instead, and "var self " is initialized to "this"
in the beginning.  This was done so that the global functions defined in this service could be referenced from the same service.  For example,
self.convertNetworkInJSONToNiceCX() calls self.setNodeAttribute() defined in the same service.  This would not work if the global functions are
defined with "this".
- Modified self.setNodeAttribute() to check if niceCX contains nodeAttributes, and if no, add it.  Also, check if
niceCX contains nodeAttributes.nodes and if no, add it.
- Added function addElementToNiceCX().
- Added function buildAttributeList() that iterates over array of Ids passed as argument, for every ID finds name of a base term with this ID,
and puts this name to a list that is returned to the caller.
- self.convertNetworkInJSONToNiceCX() modified to add "alias" and "properties" to niceCX.