This commit resolves JIRA issue NWA-53.
Modified getCitation() function to check if citation object is defined and citation['dc:identifier'] is defined.  If so,
split citation['dc:identifier'] on ":" and check if the length of resulting array (splitString) is 2.  If so, check if prefix
(splitString[0] is 'pubmed' or 'pmid').  If it is 'pubmed' or 'pmid' and splitString[1] is a number, construct a URL using
 splitString[1] as ID.  If splitString[1]  is a string and it starts with "http://" or "https://", then construct a link using
 splitString[1] .