1) createExportNetworkTask.html - only CX format can now be chosen from drop-down menu when exporting network.
2) ndex-service-factory.js - factory.getNetworkUploadURI now returns new URI for uploading network in CX format (/network/asCX);
factory.setVisibilty() now uses successHandler and errorHandler after modifying network visibility.

3) ui-service.js - task format is set to CX (for now) in createExportNetworkTask directive. submit method in
createExportNetworkTask now checks whether Network Summary, Visibility or Reference were changed, and if yes,
calls a corresponding server API.  If any of the above changes took place, we also update the view with new values (before we
called the "heavy' $route.reload method to fetch new values from server and update the view) from the modal.

4) upload-controller.js - uploader.filters.push changed to only allow CX files; uploader.onBeforeUploadItem is changed to set the
CXNetworkStream key in the form and initialize it with the file name to be uploaded.

5) upload.html - changed message to state that only CX networks are allowed for uploading.