Modified Network Tables on My Account, User, Group and Search pages to show the following images in the Status column: 1) hourglass for processing 2) green check mark for success 3) red cross for failed, and 4) exclamation mark for success with warnings.
Clicking on Error or Warning images shows a modal with error or warning(s).

Added new template, networkStatus, that describes how Status cells are to be handled.
Added new service, uiMisc, that currently has one method showNetworkWarningsOrErrors. It gets a table row and a list
of networks as arguments, finds the network from the list, extracts error message or warnings and show it in a modal.
Modififed networkName template to correctly show Network Name dependiong on the value of Network Status cell.