This resolves NWA-92 Add reference column to network list table.
1) my-account-controller.js - added columnVirtualizationThreshold:20 to $scope.networkGridOptions to make sure all columns in netwoprk table
on MyAccount page are displayed (we have more than 10 columns to display).  Added Reference column to the table, and added 4 hidden columns
to the same table (description, externalId, owenrUUID and name). These 4 columns are used as value (parameters) holders (for convenience).
Added new method getNetworkReferenceObj() extracts Reference from network property (if it exists), and constructs Reference object that is inserted
into the Reference column of the table and processed by reference.html template.

2) reference.html - template for Reference cells of Network Table.

3) showNetworkReferenceModal.html - contains code that shows modal for displaying Reference of the current network.

4) ui-services.js - added directive that pops up Modal for showing network Reference.