This commit resolves NSU-11 Downloaded network is empty.
1) getCompleteNetworkAsCX (/network/{networkid}) server API was modified to receive optional
boolean "download" argument (this API is now /network/{networkid}?download=[true|false]).
If download is true, then server sends back network as attachment (sets Content-Disposition header
in response).

2) WebApp modifications:
- download links on Group, MyAccount, Search, and User pages were modified to have ng-href pointing to
the server's /network/{networkid}?download=true. After download icon is clicked, a javasript method is called
that fills in username and password properties for the <a ...> tag for private networks for logged in users.
- download button on Network View page changed to a link.