This commit resolves NWA-133 Permission Issue and NWA-134 Removing multiple users in "manage access" misbehaves
The following changes have been made on the Manage Access page:

1) the current Admin of the network cannot be removed (the x glyph is missing);  (s)he can only be
downgraded to "Can Edit".  This is done because our server-side API can only downgrade the current admin to the Write permission,
and cannot revoke membership of the current admin.

2) UI now warns user if more than 1 admin is selected, if no admin is selected and if a different user is selected to be admin.
In the former case the warning states that a new admin is selected and the current admin will be downgraded.

3) Information on what permission is being updated/deleted is shown to the left (across) from the Discard Changes button.

4) updateNetworkMembership and deleteNetworkMembership functions are now recursive, they take as one of the argumenta a
list of permissions and call themselves until this list is empty. After operation is over, the function dispatches a certain event that is caught
by the handler that starts updateNetworkMembership function with a different list, or starts deleteNetworkMembership.
After  deleteNetworkMembership is done, we change Admin (if itnecessary), and then go back to the Network View page.

5) On the server side the update network membership and delete network membership APIs are now not-locking (they don't lock
database).  These functions are System Functions (functions that do not affect CX format).