Functionality added by this commit:
1) On Network Graphic View page:  "Add / Edit Properties" link and Edit Profile button are disabled in case the network viewed has more
than 1 subnetwork.  In this case we also show Warning message (in red) stating that "Network was exported to NDEx from Cytoscape as
collection of <number> subnetworks and cannot be edited"

2) On Network Properties Editor page: check that custom predicate doesn't collide with predefined predicate (i.e., predefined "disease"
from drop-down list of predicates and "custom... disease" will both be saved by NDEx as "disease"), so we now check for that.
We also check that customer defined predicates can only be entered once.

Custom Network Properties Editor now correctly works and saves properties for network with one or no subnetworks.