In updateVisibilityOfNetwork, we now modify visibility of networks in myAccountController.networkSearchResults as well.
Imagine the following scenario:  you created a network set, bulk selected some private networks and added them to
the newly-created set.  You try ro show case this set but get a warning that networks in this set are private and
will not be visible.  Since the networks are still selected, you can bulk-change their visibility to PUBLIC, and
show case the set again -- this time showcasing turns on since both networks in the set are PUBLIC.

This change modifies network visibility in myAccountController.networkSearchResults so that the showcasing function
could check the updated Visibility before switching showcase on.