This resolves NWA-151 bulk change visibility function doesn't apply changes to all networks.
When modifying Visibility, we now remember all network IDs of successfully modified networks.
We only send change visibility request to the server for networks that need to be changed (i.e.,
if some of the selected networks are PRIVATE and we are modifying all networks to be PRIVATE,
we only send requests for those networks that are PUBLIC).
After we have received acknowledgement from the server that the last network has been modified,
we iterate through the table of networks and modify visibilities of the networks using the
structure where we keep IDs of all successfully modified networks.

After that we modify Showcase (if required).  Again, we only modify showcase if needed,
accumulate Ids of successfully modified networks in a structure, and then update visibility of
networks in the Network Table using this structure.