This resolves NWA-148 My account table needs to auto-refresh.
Added refreshIntervalInSeconds config variable to ndex-webapp-config.js.
The value of this variable is number; it specifies how often (in seconds) the table on My Account
page refreshes.  If the value of this variable is negative, 0, or a non-number, or
if this value is not specified at all, then the auto-refresh interval is set to 30 seconds.

Removed Refresh button from the left of Tab and added Refresh Glyphicon to My Networks tab.