This commit is for NWA-179 Add Badges To NDEx UI.
It implements the following features

1) All Tasks, Sent and Received notifications in My Tasks & Notifications tab on My Networks page
are now presented in one Angular UI Grid.

2) Network Export tasks that have been created but not finished yet, are shown in bold italic.
Once a task completes or fails, it is shown in bold non-italic; this task has the "New" marking in the first column and it increases the New badge counter.  Once user clicked it, the font becomes non bold and badge counter decrements (since this task has been accessed and it is not new any more).

3) Sent network user or group access permissions, and group membership requests are shown in bold italic until
requested has been granted or denied, in which case they become "New", increment New counter shown in badge and change font to bold non-italic.  After requester has accessed/clicked it, the request is not "New" any more:
the font changes to regular and badge counter is decremented.

3) Received requests (user or group access permissions, ot group memberships requests) always stay New no matter  how many times a Recipient views them.  These requests disappear from the recipients Tasks & Notifications tab after the user acted upon them (granted or declined).

Auto- or manual refresh of My Tasks & Notifications tab pulls all Tasks, Sent and Received notifications and requests, marks them as new (if they are), populates the table and displays it.

With the current commit, we re-implement  My Tasks & Notifications functionality using Angular UI Grid. The My Tasks & Notifications tab is updated manually or via auto-refresh.