This commit is for NWA-179 Add Badges to NDEx UI.
If more than one network is selected on My Networks page, the
Bulk Actions buttons are available on the same page.
If after that we select a different tab (i.e., My Groups or My Tasks),
then the Bulk Actions buttons are not shown in these tabs any more.

In other words, the Bulk Actions for My Networks tab are now only shown
in My Networks tab.