This commit is for NWA-173 Add paging to network table in NDEx web UI.
1) Modified Bulk Delete Neworks in My Networks tab to implement the "stop and wait"
approach: the next selected network is only deleted after a success response for the
deletion of the current network has been received.

2) IN case at least one network in the selection is not owned by the user or is read-only,
the modal pops up giving exact information how many of the selected networks are read-only
and / or not owned by the user, and therefore the deletion operation cannot proceed.

3) When user selects to add the selected networks to a network set, we now correctly
show info about all network sets owned by the user (before paging, we selected all sets from the
only page we used; now we need to get all network sets regardles of on what page we currently are).