This commit is for NWA-173 Add paging to network table in NDEx web UI.
1) Modified Bulk Edit Profiles : Change Description, Change Reference, and Change Version.
This bulk operation now implements the "stop and wait"
approach: the next selected network is only updated after a success response for the
the current network has been received. In case error was received, the bulk
operation shows the error message received from the server and stops. if Cancel is hit while operation is
in progress, the operation stops, shows how many of the selected networks were processed,
 and closes modal automatically in 2 seconds.
 The operation updates all networks selected on all pages, not only on current page.
 If there are networks selected on the current page, then after operation is done
 we refresh the current page.

 2) Fixed the "Next" button on the Manage Bulk Network Access page (bulkNetworkAccessModal directive)
 to be clickable anywhere on the button, not just on the label (it is for NWA-178
 Check all buttons in WebApp and make sure they are buttons AND NOT links).