1) Modified the behavior of Bulk Actions on MyAccount page: if action is not avialable, the button gets disabled and provides explanation in a popup (html title element) why action the button is greyed out (i.e., "Unable to delete selected networks: you do not own 4 networks, 2 networks are read-only").
2) Changed the behavior of paging on My Networks tab: we do not remember selection of the previous pages.
Every time we go to a different page, the selected rows on a current/previous page get cleared.
We do remember selection of networks in My Netwrosk tab if we select a diferernt tab (i.e., My Groups or My Tasks & Notifications).

3) Important changes were made to findWhatSelectedNetworksCanBeModified(): before we had to
get network summaries FROM the server since we had selection of networks from multiple pages.
Now, since we are back to the "one-page at a time" operational mode, we do not need to query
the server since we get all information about networks on the current page when we load the current page.

4) Modified Network page to always show menu items in the More menu but disable them if actions are unavaialbe
and providing explanation why not in the popups (html title elements).