The current changes support the addition of Reference for a pre-certified networks:
1) if user requests a DOI and doesn't want to add reference later, the network becomes
"certified" (isCertified flag in network summary is set to true after
the DOI requerst is submitted). If user checks the "Let me add/modify the referecnce later"
checkbox on the DOI page of Network Properties Editor, isCertified flag is set to false.
In this case, User can select the "Add Reference" button on the Network page, add
reference and thus make the network "certified".

2) Certified networks are now marked with Gold Medal icon on My Account, User, Search
and Group pages.

3) Updated pop-out titles for Bulk Actions on My Account and on Network pages to
indicate that network(s) cannot be deleted/modified in case it is (there are)
certfied networks in the selection.