This resolves NWA-221 web app failed to load when user disallow third party cookie.
1) If third parties cookies are not allowed and "accounts.google.com" is not added to the list
of sites allowed to read/write, then the "Sign in with Google" button is greyed out and
shows the
"Google Sign In is currently unavailable because the "BLOCK THIRD-PARTY COOKIES" option is enabled in
your web browser. To use the Google Sign In feature you can do one of two things:
1. Add "accounts.google.com" to the list of websites allowed to write/read THIRD-PARTY COOKIES, or
2. Disable the "BLOCK THIRD-PARTY COOKIES" option in your browser settings." message if user
hovers over this button.

2) If third parties cookies are not allowed but "accounts.google.com" is added to the list,
user can login with the  "Sign in with Google" button.

3) If cookies are blocked by the browser, the browser presents the "Cookies are blocked by your browser.
NDEx requires that cookies be unblocked. Please unblock cookies and try again" message.