This resolves NDEX-225 Network Download Error.
ui-misc.js/self.downloadCXNetwork() was modified to build download URL of the network in the form
var link = ndexService.getNdexServerUri() + "/network/" + networkId + "?download=true";

If access key is present in the URL of the current network, then this access key is
appended to the "link" variable.

After that, we check if user logged in with google and if yes, we append add id_token
to the download URL (link); if user logged with basic auth, we prepend user name and password to the
download link.