This resolves NDEX-227 Do not show Node and Edge attributes starting with two underscores.
Added $scope.removeHiddenAttributes function that takes array of attribute names,
removes all names that start with two underscores ("__"), and returns this array back.
This function is used by $scope.getNodeAttributesNames() and $scope.getEdgeAttributesNames()
to remove the "hidden" names.

populateNodeTable(), populateEddeTable(), refreshEdgeTable() and refreshNodeTable() now
check if any of the attributes start with "__" and do not add these columns to the tables.