This resolves NWA-232 Network View Page needs to refresh after DOI request is submitted.
1) removed unused controller edit-network-properties-controller.js and
form editNetworkProperties.html (they were replaced by edit-network-properties-controller-fixed-form.js and
views/editNetworkPropertiesFixedForm.html).

2) editor.save in edit-network-properties-controller.js now takes 2 arguments: boolean indicator
that shows whether to request a DOI and info block to send in email if DOI is requested.  These two arguments are
false and null if the function is called from the view; but true and not null in case it is called by
editor.saveAndSubmitDOI(). editor.save now requests DOI after the call to ndexService.setNetworkSummaryV2() has returned.

3) also, modified networkController.getStatusOfShareableURL() in network-controller.js to check if Cytoscape is running
 and its' version in case user is the owner of the private network and it doesn't have an access key.