1) Use a newer version of font-awesome in bower.json.  Version 4.7.0 includes the "Clone" icon (<i class="fa fa-clone"></i>) that we now use instead of "Clone" button on the Network page.
2) cytoscape_off.png image is removed.  cytoscape_on.png is renamed to cytoscape_logo.png which
is now styled in ndex-angular-site.css to be both in "on" and "off" states.

3) behavior of "Open in Cytoscape" feature is changed:  if it is on and selected,
the title of the button changes to "Opening <network name> in Cytoscape".  Once the call to cyREST.exportNetworkToCytoscape
sucessfully returns, the titlw of the button chanhges to "Opened" and then after 2 secpnds in changes
back to "Open in Cytoscape".

4) All buttons, links and drop-up menu items now use the title tooltips.