This resolves NWA-165 image link is broken in the down for maintenance page.
In main.js, if we check for the server and server is down, we hide
NDEx logo shown in the top left corner.

In index.html, in case server is down, we do not show the Collapsed menu button on the right.

On a different subject, in network.html, renamed the "Log in to operate on this network" button
to "Log in".  Added pop-up title to this button.  If the network is a Cytoscape
Collection, the title just states "Log in"; otherwise, it states
"Log in to operate on this network".