This resolves SRV2-80 red error box for query results too large.
1) In case server returns error for Neighborhood Query or Advanced Query, check for string "EdgeLimitExceeded"
in the error.data.message.  If "EdgeLimitExceeded" is present, then treat the response as not
an error but as warning.

2) Modified the behavior of Neighb. and Advanced queries:  if query returns
error or warning, then we just show this warning or error in the bottom part
of Cytoscape Canvas or Table - we do not re-draw the network (re-populate the table).
The Query box stays avaialable in case user wants to issue a differentr query.
Only if query is successful, we present a sub-network screen where user can
save the query or go back to the initial network.