This resolves NWA-235 This network is not rendered correctly in NDEx.
Networks received fomr Cytoscape can have "list of ... strings/boolens/numbers"
types for their nodes/edges definitions.

To resolves this issue, we added handling of these types the same way Cytoscape does:

- Get the first element from the list of elements.  Attributes of node or edge of this first element will be
used to style the node/edge.  This behavior simulates the behavior of Cytoscape:

1) if a node column has the type "list of ... (strings/numbers/boolens), then Cytoscape uses the first element
in the list to style the node (or edge).
For example, if a column has the type ['protein', 'gene'] (list of strings), then Cytoscape uses the shape of
'protein' (the first element in the list) to draw this node (the shape of 'protein' in Cytoscape is 'round rectangle').

2) If, however, a node has the list ['gene', 'protein'], then Cytoscape uses the shape of
'gene' (shape of 'gene' is 'diamond') to draw this node.

Thus, we added a new function that extracts the first element from the loist of elements.