Added Full Index Advisory modal to the Network Properties Editor: if user selects/clicks the Full Index check box, a modal pops with the following warning:
Full Index is resource-expensive feature and should only be used for high quality,
published networks made available to the scientific community... Although we
understand that this feature might also be useful for PRIVATE networks,
we ask you to be considerate and use it sparingly.

<strong>Would you like to activate the Full Index feature?<strong>

If user selects "Proceed", the check box get checked.