Replaced Query Depth 3-step with 1-Step Interconnect.
When running neigborhood and advanced queries, we now save/keep the original network in NiceCX and
then restore it (do no fetch it again from the server) when hitting the Go Back To Original network button.

Every time we run query, we now create a new instance of Spinner and destroy it
after we go back to the original network.

Fixed bug where you enter a lot of spaces on the Query string and run query.
We now handle this cas and give the "Please fill out this field" warning.