This resolves NWA-254 Layout not displayed correctly.
The problem was that the original network has two nodes with the same name
and exactly the same coordinates.  The cyService.allNodesHaveUniquePositions() detected
that two nodes had the same coordinates, returned false, and that caused the
'cose' instead of 'preset' layout to be used.

We removed the check for uniqueness of coordinates of all nodes.
We now check if network (or network sample) in drawCXNetworkOnCanvas has
laoyout and if yes then use 'preset', use 'cose' otherwise:

var layoutName = (cxNetwork.cartesianLayout) ? 'preset' : 'cose';