This resolves 1) NWA-256 Query result in graphic view doesn't display node or edge info when click on them, and 2) NWA-257 In Firefox, cannot close Advanced Query.
Modified code to initialize originalNiceCX property of networkService in networkServices.js
when retrieving an entire or sample network in NiceCX with getCompleteNetworkInCXV2() or in getNetworkSampleV2().
The originalNiceCX is used to remember the original network that is displayed when user
comes to the Network page.  We use originalNiceCX to restore the original network
after we get back to the Network Page from Neighborhood or Advanced query
by clicking the "Back to Original Network" button.

The currentNiceCX keeps the network currently displayed in the Graph or Network view.
The currentNiceCX keeps either the original network in nice CX (in which case it has the same
contents as originalNiceCX), or the result of Advanced/Neigborhood query (in which
case its contents differs from that of originalNiceCX).
Selecting the ""Back to Original Network" restores the originalNiceCX in the currentNiceCX.

In networkController.backToOriginalNetwork(), when getting back to the Network View from
Advanced Query, we need to check if event is undefined (it is undefined for FireFox, and
defined for Chrome, Safari and Outlook).