Augmented the Neighborhood Query :
1) when running neigborhood query, if server returns more than configurable parameter
networkQueryLimitLow, we present user with a choice of Cancel, Save the Current Query Result
or Keep Running the query.

2) if user select Keep running, then if network returned result with number of edges less than
configurable networkQueryLimitHigh, this result is presented in Cytoscape.js.
Otherwise, warning is shown stating that the network query result is too big, proceed?
If user selects to proceed, we run the query and silently save its results to users' account.

This code doesn't notify about finishing querying and it doesn't support yet anonymous users.