Redesigned execution of Neighborhood Query:
1) removed networkQueryLimitLow and networkQueryLimitHigh from ndex-webapp-config.js;
replaced them with networkQueryEdgeLimit (default is 50000)

2) introduced two hard-coded parameters to network-controller.js:
networkController.queryEdgeLimitToShowGraph=1000 and
networkController.queryEdgeLimitToShowTableOnly=3000.

3) if neighborhood query returned no more than 1000 edges, both Graph and
Table views are available for viewing the query result (for both logged in
and anonymous users)

4) if neighborhood query returned more than 1000 but no more than 3000 edges,
we show the result in Table but Graph view is still available. Trying
to Switch to Graph view will show warning stating that browser might be slow in the
Graph View (for both logged in and anonymous users).

5) if neighborhood query returned more than 3000 edges, Graph mode is unavailable
 (for both logged in and anonymous users).

6) if neighborhood query returned more than networkQueryEdgeLimit edges, Graph mode is unavailable,
and we prompt a logged in user to run the query in the background and save it.
An anonymous user is suggested to log in in order to save the query.