Thi resolves NWA-277 'Label position is not rendered correctly when used in combination with a discrete mapping in Cytoscape'.
I added 'NODE_LABEL_POSITION' property to visualPropertyMap structure.

Function getCyVisualAttributeValue() now checks if cyVisualAttributeType is a Node Label pPosition, and
if it is, getNodeLabelPosition() is called to build a map of coordinates
in the form {'text-halign': <value1>, 'text-valign': <value2>} and return it to the caller.

discreteMappingStyle() checks if cyVisualAttribute is a Node Label Position and
if it is, cyVisualAttributePair['text-valign'] and cyVisualAttributePair['text-halign']
are initialized with the values returned by getCyVisualAttributeValue() - actual
postion of the Node Label that will be used to render the node.