This implements the following changes from Preserving Tables Customization in NDEx WebApp design document (https://docs.google.com/document/d/1beyvmW76zFGtsiTqyQx2rAjFooG4_O0vX1vy_hzgi1g/edit#heading=h.v2cy4bj1w2uh):
- routing to MyAccount page now includes page number in the form /myAccount/p<number>.
The very first page indicates no page number an is shown in URL as /myAccount.
So now when flipping through network pages in My Networks table, the URL shows page number.

- hitting browser Back button shows previous screen (portion) of networks in My Network table
(works same as hitting back button of My Networks table, but reloads the entire page and not
only the network portion).

- implemented remembering sorting of pages and tables page size.

- important change: in my-account-controller, every time we arrive to a new page
in My Networks table, we need not only to repopulate the page, but redefine it first to
apply sorting that was created by user earlier.  Once network header is defined, it
 is hard to programmatically change its' sorting order.